requirements.txt
fastparquet==2024.5.0
hurry.filesize==0.9
jinja2==3.1.4
jupyterlab==4.3.0
numpy==2.1.2
omdbapi==0.7.0
pandas==2.2.3
pyarrow==18.0.0
pyspark==3.5.3
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
tabulate==0.9.0